+++ Doctrine/Orm/Filter/SearchFilter.php
@@ -15,6 +15,7 @@ use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\ORM\QueryBuilder;
 use Dunglas\ApiBundle\Api\IriConverterInterface;
 use Dunglas\ApiBundle\Api\ResourceInterface;
+use Dunglas\ApiBundle\Doctrine\Orm\Util\QueryUtils;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 
@@ -78,42 +79,70 @@ class SearchFilter extends AbstractFilter
             return;
         }
 
-        $metadata = $this->getClassMetadata($resource);
-        $fieldNames = array_flip($metadata->getFieldNames());
-
         foreach ($this->extractProperties($request) as $property => $value) {
-            if (null === $value || !$this->isPropertyEnabled($property)) {
+            if (
+                !$this->isPropertyEnabled($property)
+                || !$this->isPropertyMapped($property, $resource, true)
+                || null === $value
+            ) {
                 continue;
             }
 
-            if (isset($fieldNames[$property])) {
+            $alias = 'o';
+            $field = $property;
+
+            if ($this->isPropertyNested($property)) {
+                $propertyParts = $this->splitPropertyParts($property);
+
+                $parentAlias = $alias;
+
+                foreach ($propertyParts['associations'] as $association) {
+                    $alias = QueryUtils::generateJoinAlias($association);
+                    $queryBuilder->join(sprintf('%s.%s', $parentAlias, $association), $alias);
+                    $parentAlias = $alias;
+                }
+
+                $field = $propertyParts['field'];
+
+                $metadata = $this->getNestedMetadata($resource, $propertyParts['associations']);
+            } else {
+                $metadata = $this->getClassMetadata($resource);
+            }
+
+            if ($metadata->hasField($field)) {
                 if (!is_string($value)) {
                     continue;
                 }
 
                 $partial = null !== $this->properties && self::STRATEGY_PARTIAL === $this->properties[$property];
 
-                if ('id' === $property) {
+                if ('id' === $field) {
                     $value = $this->getFilterValueFromUrl($value);
                 }
 
+                $valueParameter = QueryUtils::generateParameterName($field);
+
                 $queryBuilder
-                    ->andWhere(sprintf('o.%1$s LIKE :%1$s', $property))
-                    ->setParameter($property, $partial ? sprintf('%%%s%%', $value) : $value)
+                    ->andWhere(sprintf('%s.%s LIKE :%s', $alias, $field, $valueParameter))
+                    ->setParameter($valueParameter, $partial ? sprintf('%%%s%%', $value) : $value)
                 ;
-            } elseif ($metadata->isSingleValuedAssociation($property)) {
+            } elseif ($metadata->isSingleValuedAssociation($field)) {
                 if (!is_string($value)) {
                     continue;
                 }
 
                 $value = $this->getFilterValueFromUrl($value);
 
+                $association = $field;
+                $associationAlias = QueryUtils::generateJoinAlias($association);
+                $valueParameter = QueryUtils::generateParameterName($association);
+
                 $queryBuilder
-                    ->join(sprintf('o.%s', $property), 'api_'.$property)
-                    ->andWhere(sprintf('api_%1$s.id = :%1$s', $property))
-                    ->setParameter($property, $value)
+                    ->join(sprintf('%s.%s', $alias, $association), $associationAlias)
+                    ->andWhere(sprintf('%s.id = :%s', $associationAlias, $valueParameter))
+                    ->setParameter($valueParameter, $value)
                 ;
-            } elseif ($metadata->isCollectionValuedAssociation($property)) {
+            } elseif ($metadata->isCollectionValuedAssociation($field)) {
                 $values = $value;
                 if (!is_array($values)) {
                     $values = [$value];
@@ -130,10 +159,14 @@ class SearchFilter extends AbstractFilter
 
                 $values = array_map([$this, 'getFilterValueFromUrl'], $values);
 
+                $association = $field;
+                $associationAlias = QueryUtils::generateJoinAlias($association);
+                $valuesParameter = QueryUtils::generateParameterName($association);
+
                 $queryBuilder
-                    ->join(sprintf('o.%s', $property), 'api_'.$property)
-                    ->andWhere(sprintf('api_%1$s.id IN (:%1$s)', $property))
-                    ->setParameter($property, $values)
+                    ->join(sprintf('%s.%s', $alias, $association), $associationAlias)
+                    ->andWhere(sprintf('%s.id IN (:%s)', $associationAlias, $valuesParameter))
+                    ->setParameter($valuesParameter, $values)
                 ;
             }
         }
@@ -145,28 +178,44 @@ class SearchFilter extends AbstractFilter
     public function getDescription(ResourceInterface $resource)
     {
         $description = [];
-        $metadata = $this->getClassMetadata($resource);
 
-        foreach ($metadata->getFieldNames() as $fieldName) {
-            $found = isset($this->properties[$fieldName]);
-            if ($found || null === $this->properties) {
-                $description[$fieldName] = [
-                    'property' => $fieldName,
-                    'type' => $metadata->getTypeOfField($fieldName),
-                    'required' => false,
-                    'strategy' => $found ? $this->properties[$fieldName] : self::STRATEGY_EXACT,
-                ];
-            }
+        $properties = $this->properties;
+        if (null === $properties) {
+            $properties = array_fill_keys($this->getClassMetadata($resource)->getFieldNames(), null);
         }
 
-        foreach ($metadata->getAssociationNames() as $associationName) {
-            if ($this->isPropertyEnabled($associationName)) {
-                $paramName = $associationName;
-                if ($metadata->isCollectionValuedAssociation($associationName)) {
+        foreach ($properties as $property => $strategy) {
+            if (!$this->isPropertyMapped($property, $resource, true)) {
+                continue;
+            }
+
+            if ($this->isPropertyNested($property)) {
+                $propertyParts = $this->splitPropertyParts($property);
+
+                $field = $propertyParts['field'];
+
+                $metadata = $this->getNestedMetadata($resource, $propertyParts['associations']);
+            } else {
+                $field = $property;
+
+                $metadata = $this->getClassMetadata($resource);
+            }
+
+            if ($metadata->hasField($field)) {
+                $description[$property] = [
+                    'property' => $property,
+                    'type' => $metadata->getTypeOfField($field),
+                    'required' => false,
+                    'strategy' => isset($this->properties[$property]) ? $this->properties[$property] : self::STRATEGY_EXACT,
+                ];
+            } elseif ($metadata->hasAssociation($field)) {
+                $association = $field;
+                $paramName = $property;
+                if ($metadata->isCollectionValuedAssociation($association)) {
                     $paramName .= '[]';
                 }
                 $description[$paramName] = [
-                    'property' => $associationName,
+                    'property' => $property,
                     'type' => 'iri',
                     'required' => false,
                     'strategy' => self::STRATEGY_EXACT,